feat(wip): commander実装中

+++ broccoli-nandio-rp2040/src/driver.rs
@@ -58,7 +58,7 @@ impl Driver for Rp2040FwDriver<'_> {
     }
 
     /// Reset NAND IC
-    fn reset(&mut self, cs_index: u32) {
+    fn reset(&mut self, cs_index: usize) {
         self.nandio_pins.assert_cs(cs_index);
         self.nandio_pins.input_command(CommandId::Reset as u8, || {
             self.delay.delay_us(DELAY_US_FOR_COMMAND_LATCH)
@@ -69,7 +69,7 @@ impl Driver for Rp2040FwDriver<'_> {
     }
 
     /// Read NAND IC ID
-    fn read_id(&mut self, cs_index: u32) -> (bool, [u8; ID_READ_CMD_BYTES]) {
+    fn read_id(&mut self, cs_index: usize) -> (bool, [u8; ID_READ_CMD_BYTES]) {
         let mut id_read_results = [0x00, 0x00, 0x00, 0x00, 0x00];
 
         self.nandio_pins.assert_cs(cs_index);
@@ -96,7 +96,7 @@ impl Driver for Rp2040FwDriver<'_> {
     }
 
     /// Read NAND IC status
-    fn read_status(&mut self, cs_index: u32) -> StatusOutput {
+    fn read_status(&mut self, cs_index: usize) -> StatusOutput {
         let mut status = [0x00];
 
         self.nandio_pins.assert_cs(cs_index);
@@ -116,7 +116,7 @@ impl Driver for Rp2040FwDriver<'_> {
     /// Read NAND IC data
     fn read_data(
         &mut self,
-        cs_index: u32,
+        cs_index: usize,
         address: Address,
         read_data_ref: &mut [u8],
         read_bytes: usize,
@@ -162,14 +162,14 @@ impl Driver for Rp2040FwDriver<'_> {
 
     fn read_id_async(
         &mut self,
-        cs_index: u32,
+        cs_index: usize,
     ) -> impl Future<Output = (bool, [u8; ID_READ_CMD_BYTES])> {
         async move { self.read_id(cs_index) }
     }
 
     fn read_data_async(
         &mut self,
-        cs_index: u32,
+        cs_index: usize,
         address: Address,
         read_data_ref: &mut [u8],
         read_bytes: usize,
@@ -177,7 +177,7 @@ impl Driver for Rp2040FwDriver<'_> {
         async move { self.read_data(cs_index, address, read_data_ref, read_bytes) }
     }
 
-    fn erase_block(&mut self, cs_index: u32, address: Address) -> Result<StatusOutput, Error> {
+    fn erase_block(&mut self, cs_index: usize, address: Address) -> Result<StatusOutput, Error> {
         self.nandio_pins.assert_cs(cs_index);
         self.nandio_pins
             .input_command(CommandId::AutoBlockEraseFirst as u8, || {
@@ -229,7 +229,7 @@ impl Driver for Rp2040FwDriver<'_> {
 
     fn write_data(
         &mut self,
-        cs_index: u32,
+        cs_index: usize,
         address: Address,
         write_data_ref: &[u8],
         write_bytes: usize,
@@ -291,11 +291,11 @@ impl Driver for Rp2040FwDriver<'_> {
         async { self.init_pins() }
     }
 
-    fn reset_async(&mut self, cs_index: u32) -> impl Future<Output = ()> {
+    fn reset_async(&mut self, cs_index: usize) -> impl Future<Output = ()> {
         async move { self.reset(cs_index) }
     }
 
-    fn read_status_async(&mut self, cs_index: u32) -> impl Future<Output = StatusOutput> {
+    fn read_status_async(&mut self, cs_index: usize) -> impl Future<Output = StatusOutput> {
         async move { self.read_status(cs_index) }
     }
 
@@ -305,7 +305,7 @@ impl Driver for Rp2040FwDriver<'_> {
 
     fn erase_block_async(
         &mut self,
-        cs_index: u32,
+        cs_index: usize,
         address: Address,
     ) -> impl Future<Output = Result<StatusOutput, Error>> {
         async move { self.erase_block(cs_index, address) }
@@ -313,7 +313,7 @@ impl Driver for Rp2040FwDriver<'_> {
 
     fn write_data_async(
         &mut self,
-        cs_index: u32,
+        cs_index: usize,
         address: Address,
         write_data_ref: &[u8],
         write_bytes: usize,
