+++ broccoli-nandio-rp2040/src/pins.rs
@@ -258,7 +258,7 @@ impl NandIoPins<'_> {
 
     /// set CS
     /// cs_index: chip select index
-    pub fn assert_cs(&mut self, cs_index: u32) {
+    pub fn assert_cs(&mut self, cs_index: usize) {
         match cs_index {
             0 => {
                 self.ceb0.set_state(bsp::hal::gpio::PinState::Low).unwrap();
