+++ broccoli-nandio/src/driver.rs
@@ -138,52 +138,52 @@ pub trait Driver {
     fn set_write_protect_async(&mut self, enable: bool) -> impl Future<Output = ()>;
 
     /// Reset NAND IC
-    fn reset(&mut self, cs_index: u32);
-    fn reset_async(&mut self, cs_index: u32) -> impl Future<Output = ()>;
+    fn reset(&mut self, cs_index: usize);
+    fn reset_async(&mut self, cs_index: usize) -> impl Future<Output = ()>;
 
     /// Read NAND IC ID
-    fn read_id(&mut self, cs_index: u32) -> (bool, [u8; ID_READ_CMD_BYTES]);
-    fn read_id_async(&mut self, cs_index: u32) -> impl Future<Output = (bool, [u8; 5])>;
+    fn read_id(&mut self, cs_index: usize) -> (bool, [u8; ID_READ_CMD_BYTES]);
+    fn read_id_async(&mut self, cs_index: usize) -> impl Future<Output = (bool, [u8; 5])>;
 
     /// Read NAND IC data
     fn read_data(
         &mut self,
-        cs_index: u32,
+        cs_index: usize,
         address: Address,
         read_data_ref: &mut [u8],
         read_bytes: usize,
     ) -> Result<(), Error>;
     fn read_data_async(
         &mut self,
-        cs_index: u32,
+        cs_index: usize,
         address: Address,
         read_data_ref: &mut [u8],
         read_bytes: usize,
     ) -> impl Future<Output = Result<(), Error>>;
 
     /// Read NAND IC status
-    fn read_status(&mut self, cs_index: u32) -> StatusOutput;
-    fn read_status_async(&mut self, cs_index: u32) -> impl Future<Output = StatusOutput>;
+    fn read_status(&mut self, cs_index: usize) -> StatusOutput;
+    fn read_status_async(&mut self, cs_index: usize) -> impl Future<Output = StatusOutput>;
 
     /// Erase NAND IC block
-    fn erase_block(&mut self, cs_index: u32, address: Address) -> Result<StatusOutput, Error>;
+    fn erase_block(&mut self, cs_index: usize, address: Address) -> Result<StatusOutput, Error>;
     fn erase_block_async(
         &mut self,
-        cs_index: u32,
+        cs_index: usize,
         address: Address,
     ) -> impl Future<Output = Result<StatusOutput, Error>>;
 
     /// Write NAND IC data
     fn write_data(
         &mut self,
-        cs_index: u32,
+        cs_index: usize,
         address: Address,
         write_data_ref: &[u8],
         write_bytes: usize,
     ) -> Result<StatusOutput, Error>;
     fn write_data_async(
         &mut self,
-        cs_index: u32,
+        cs_index: usize,
         address: Address,
         write_data_ref: &[u8],
         write_bytes: usize,
